scripts/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.30.0
pinecone-client>=3.0.0
python-dotenv>=1.0.0
